requirements.txt
streamlit
google-generativeai
pandas
Pillow
python-dotenv
openpyxl
